Prevents seed_data from being run in production

+++ spec/data/import.js
@@ -6,6 +6,11 @@ const usersImporter = require('./importUsers');
 
 const run = async (cb) => {
   try {
+
+    if (process.env.NODE_ENV === 'production') {
+      throw new Error('Cannot import data in production mode.');
+    }
+
     await mongoose.connection.dropDatabase()
     await businessesImporter.run();
     await categoriesImporter.run();
@@ -13,10 +18,11 @@ const run = async (cb) => {
     await usersImporter.run();
 
     if (cb !== undefined) {
-      cb();
+      cb(true);
     }
   } catch(e) {
     console.log(e);
+    cb(false);
   }
 };
